Refactor code creating library for snake functions.
