+++ templates/index.html
@@ -3,7 +3,7 @@
 {% block links %}
 
     <link rel="stylesheet" href="../static/index.css">
-    <script src="../static/snake.js"></script>
+    <script type="module" src="../static/script.js"></script>
 
 {% endblock %}
